Add aws delete to destroy project
